INIT: Semi-Successful Launch
This commit is the initial commit to this repo. It includes an updated README that goes through the primary things learned up until now and gives other documentation. It includes an EC2 directory that will act as a log/reference on how to provision an EC2 instance using Terraform via AWS provider. Furthermore, the main.tf file contains code that is actively setting up a Classic Load Balancer, two EC2 instances with two seperate security groups via AWS, using the domain judahrjackson.com with automatic DNS record keeping with Cloudflare for the load balancer. The reason this commit is only considered a Semi-Successful Launch is because it is a consistently reproducable 522 (Connection timed out) error as opposed to a 1016 (Domain not reachable) error. Basically, right now the browser is working, the domain routing through Cloudflare is working, but the host (judahrjackson.com) is having an error. From what I can tell right now, Cloudflare connects to the LB, which forwards to an EC2 instance (the backend) and the EC2 instance serves HTTP via port 80 or HTTPS via port 443, and the problem right now is that the AWS LB forwarding an EC2 instance isn't happening.