Load Balancer Update
Instances were responding via `curl http://<public ip address>` (when cidr was opened up to anyone), and load balancer was responding with responses from both instances using both the elb dns name and direct link to judahrjackson.com via curl method. Meaning, the load balancer was up and running and the routing through Cloudflare was working fine, but the browser was still getting the 522 Error. It was deduced that the load balancer's security group was not allowing both ports 80 and 443 inbound. This is important because Cloudflare tries to connect to the ELB on port 443 (HTTPS). Once this was added all the previous checks passed and the browser no longer received the 522 error. Instances security groups must have their traffic limited again, but for now (for testing/debugging) they are left open.